requirements.txt
streamlit==1.51.0
transformers==4.57.1
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
numpy==2.3.4
pandas==2.2.3
altair==5.5.0
tqdm==4.67.1
huggingface-hub>=0.34.0
regex==2025.11.3
fsspec==2025.10.0
typing_extensions==4.12.2
protobuf==5.28.3
watchdog==6.0.0
click==8.3.0
httpx==0.28.1
